Event class improvements; implement FontFactory; Refactored Animation classes; Add more methods for Pulse; some size parameters changed from int to size_t

+++ Engine/Include/Graphics/Resource/FontAtlas.h
@@ -20,7 +20,7 @@ namespace engine
 				FontAtlas(std::unique_ptr<engine::graphics::resource::ISpriteAtlas> spriteAtlas);
 				virtual ~FontAtlas() = default;
 
-				bool Initialize(const std::string& fontName = "Arial", const unsigned int fontSize = 12) override final;
+				bool Initialize(const std::string& fontName = "Arial", const size_t fontSize = 12) override final;
 
 				void Reset() override final;
